refactor(playlist-api): moved playlist manager window to Playlist.kt
