+++ src/main/xerus/monstercat/tabs/TabPlaylist.kt
@@ -83,179 +83,8 @@ class TabPlaylist : VTab() {
 	}
 	
 	init {
-		addButton("Playlists from Monstercat.com..."){ playlistDialog() }
+		addButton("Playlists from Monstercat.com..."){ PlaylistManager.playlistDialog() }
 		fill(table)
-	}
-	
-	private suspend fun loadPlaylist(apiConnection: APIConnection){
-		val tracks = apiConnection.getTracks()
-		tracks?.forEachIndexed { index, track ->
-			val found = Cache.getTracks().find {
-				it.id == track.id
-			}
-			if (found != null)
-				tracks[index] = found
-			else
-				tracks.removeAt(index)
-		}
-		if (tracks != null && tracks.isNotEmpty()) {
-			Player.reset()
-			Playlist.setTracks(tracks)
-		}
-	}
-
-	private fun playlistDialog(){
-		val connection = APIConnection("api", "playlist").fields(ConnectPlaylist::class)
-
-		val parent = VBox()
-		val stage = App.stage.createStage("Monstercat.com Playlists", parent)
-		stage.initModality(Modality.WINDOW_MODAL)
-		val connectTable = TableView<ConnectPlaylist>()
-
-		// Common playlist functions
-		fun load(runAfter: () -> Unit = {}) {
-			if (connectTable.selectionModel.selectedItem != null) {
-				GlobalScope.async {
-					loadPlaylist(APIConnection("api", "playlist", connectTable.selectionModel.selectedItem.id, "tracks"))
-					onFx { runAfter.invoke() }
-				}
-			}
-		}
-		fun loadUrl(runAfter: () -> Unit = {}) {
-			val subParent = VBox()
-			val subStage = stage.createStage("Load from URL", subParent)
-			subStage.initModality(Modality.WINDOW_MODAL)
-			val textField = TextField().apply { promptText = "URL" }
-			subParent.add(textField)
-			subParent.addRow(createButton("Load"){
-				val playlistId = textField.text.substringAfterLast("/")
-				if (playlistId.length == 24){
-					GlobalScope.async {
-						try {
-							loadPlaylist(APIConnection("api", "playlist", playlistId, "tracks"))
-							onFx { subStage.close(); runAfter.invoke() }
-						}catch (e: Exception){
-							onFx {
-								monsterUtilities.showAlert(Alert.AlertType.WARNING, "No playlist found", content = "No playlist were found at ${textField.text}.")
-							}
-						}
-					}
-				}else{
-					monsterUtilities.showAlert(Alert.AlertType.WARNING, "Playlist URL invalid", content = "${textField.text} is not a valid URL.")
-				}
-			}, createButton("Cancel"){
-				subStage.close()
-			})
-			subStage.show()
-		}
-		fun replace(runAfter: () -> Unit = {}) {
-			if (connectTable.selectionModel.selectedItem != null) {
-				GlobalScope.async {
-					APIConnection("v2", "playlist", connectTable.selectionModel.selectedItem.id).editPlaylist(Playlist.tracks)
-					onFx { runAfter.invoke() }
-				}
-			}
-		}
-		fun delete(runAfter: () -> Unit = {}) {
-			if (connectTable.selectionModel.selectedItem != null) {
-				GlobalScope.async {
-					APIConnection("v2", "playlist", connectTable.selectionModel.selectedItem.id).editPlaylist(deleted = true)
-					onFx { runAfter.invoke() }
-				}
-			}
-		}
-		fun new(runAfter: () -> Unit = {}) {
-			val subParent = VBox()
-			val subStage = stage.createStage("New Playlist", subParent)
-			subStage.initModality(Modality.WINDOW_MODAL)
-			val textField = TextField().apply { promptText = "Name" }
-			val publicTick = CheckBox("Public")
-			subParent.children.addAll(textField, publicTick)
-			subParent.addRow(createButton("Create"){
-				GlobalScope.async {
-					APIConnection("api", "playlist").createPlaylist(textField.text.let { if (it.isBlank()) "New Playlist" else it }, Playlist.tracks, publicTick.isSelected)
-					onFx { subStage.close(); runAfter.invoke() }
-				}
-			}, createButton("Cancel"){
-				subStage.close()
-			})
-			subStage.show()
-		}
-		fun rename(runAfter: () -> Unit = {}) {
-			val subParent = VBox()
-			val subStage = stage.createStage("Rename Playlist", subParent)
-			subStage.initModality(Modality.WINDOW_MODAL)
-			val textField = TextField().apply { promptText = "Name" }
-			subParent.add(textField)
-			subParent.addRow(createButton("Create"){
-				if (connectTable.selectionModel.selectedItem != null) {
-					GlobalScope.async {
-						APIConnection("v2", "playlist", connectTable.selectionModel.selectedItem.id).editPlaylist(name = textField.text.let { if (it.isBlank()) "Unnamed" else it })
-						onFx { subStage.close(); runAfter.invoke() }
-					}
-				}
-			}, createButton("Cancel"){
-				subStage.close()
-			})
-			subStage.show()
-		}
-
-		val playlists = FXCollections.observableArrayList<ConnectPlaylist>()
-		fun updatePlaylists() {
-			playlists.clear()
-			if (APIConnection.connectValidity.value != ConnectValidity.NOGOLD && APIConnection.connectValidity.value != ConnectValidity.GOLD){
-				connectTable.placeholder = Label("Please connect using connect.sid in the downloader tab.")
-			}else{
-				connectTable.placeholder = Label("Loading...")
-				GlobalScope.async {
-					val results = connection.getPlaylists()
-					if (results != null && results.isNotEmpty())
-						playlists.addAll(results)
-					else
-						onFx {
-							connectTable.placeholder = Label("No playlists were found on your account.")
-						}
-				}
-			}
-		}
-
-		connectTable.apply {
-
-			columnResizePolicy = TableView.CONSTRAINED_RESIZE_POLICY
-			columns.addAll(TableColumn<ConnectPlaylist, String>("Name").apply {
-				cellValueFactory = Callback<TableColumn.CellDataFeatures<ConnectPlaylist, String>, ObservableValue<String>> { p ->
-					ImmutableObservable(p.value.name)
-				}
-			}, TableColumn<ConnectPlaylist, String>("Size").apply {
-				cellValueFactory = Callback<TableColumn.CellDataFeatures<ConnectPlaylist, String>, ObservableValue<String>> { p ->
-					ImmutableObservable(p.value.tracks.size.toString())
-				}
-			})
-
-			items = playlists
-
-			selectionModel.selectionMode = SelectionMode.SINGLE
-
-			setOnMouseClicked { if (it.button == MouseButton.PRIMARY && it.clickCount == 2) load() }
-
-			val publicMenuItem = CheckMenuItem("Public", {
-				if (connectTable.selectionModel.selectedItem != null) {
-					GlobalScope.async {
-						APIConnection("api", "playlist", connectTable.selectionModel.selectedItem.id).editPlaylist(public = it)
-						onFx { updatePlaylists() }
-					}
-				}
-			})
-			contextMenu = ContextMenu(publicMenuItem, SeparatorMenuItem(), MenuItem("Save into") { replace { updatePlaylists() }; }, MenuItem("Rename playlist") { rename { updatePlaylists() } }, MenuItem("Delete playlist") { delete { updatePlaylists() } })
-			contextMenu.setOnShown { publicMenuItem.isSelected = selectionModel.selectedItem?.public ?: false }
-
-			updatePlaylists()
-		}
-
-		parent.add(Label("Tip : You can right-click a playlist to edit it without the window closing each time !"))
-		parent.addRow(createButton("Load"){ load(); stage.close() }, createButton("From URL..."){ loadUrl { stage.close() } }, createButton("Save into selected"){ replace { stage.close() } }, createButton("Save as new..."){ new { stage.close() } })
-		parent.fill(connectTable, 0)
-		stage.show()
 	}
 
 	private val selectedTrack: Track
